Добавляет реализацию редакитирования изображения
