+++ index.html
@@ -5,6 +5,7 @@
   <meta name="viewport" content="width=device-width,initial-scale=1">
   <link rel="stylesheet" href="css/normalize.css">
   <link rel="stylesheet" href="css/style.css">
+  <link href="vendor/nouislider/nouislider.css" rel="stylesheet">
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
   <title>Кекстаграм</title>
 </head>
@@ -226,6 +227,7 @@
       </div>
     </section>
   </template>
+<script src="vendor/nouislider/nouislider.js"></script>
 <script src="vendor/pristine/pristine.min.js"></script>
 <script  src="js/main.js" type="module"></script>
 </body>
