+++ js/pictures/render-miniatures.js
